增加广告设置中检测U盘部分功能

+++ app/src/main/java/cn/edu/stu/max/cocovendor/Activity/AdSettingActivity.java
@@ -11,6 +11,7 @@ import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.ListView;
+import android.widget.TextView;
 import android.widget.Toast;
 import android.widget.VideoView;
 
@@ -22,6 +23,7 @@ import java.util.Map;
 
 import cn.edu.stu.max.cocovendor.JavaClass.FileService;
 import cn.edu.stu.max.cocovendor.JavaClass.MyInternalListAdapter;
+import cn.edu.stu.max.cocovendor.JavaClass.MySettingListAdapter;
 import cn.edu.stu.max.cocovendor.JavaClass.MyUSBListAdapter;
 import cn.edu.stu.max.cocovendor.R;
 
@@ -30,6 +32,8 @@ public class AdSettingActivity extends Activity {
 
     private final static String FROMPATH = "/mnt/usb_storage/USB_DISK2/udisk0/advertisement/";   // U盘广告存储路径
     private final static String TOPATH = "/storage/sdcard0/tencent/QQfile_recv/b/";               // 本机广告存储路径
+    private final static int FROMPATHFLAG = 0;          // 用来判断U盘广告存储路径
+    private final static int TOPATHFLAG = 1;            // 用来判断本机广告存储路径
 
     private VideoView videoViewAd;    // 播放广告视频
 
@@ -37,23 +41,33 @@ public class AdSettingActivity extends Activity {
 
     private ListView listViewAdUSB;         // U盘广告列表
     private ListView listViewAdInternal;   // 本机广告列表
+    private ListView listViewAdSetting;    // 本机广告设置列表
 
     private MyUSBListAdapter myUSBListAdapter;    // U盘广告内容适配器
     private MyInternalListAdapter myInternalListAdapter;    // 本机广告内容适配器
+    private MySettingListAdapter mySettingListAdapter;      // 本机广告内容设置适配器
 
     private boolean hasAddedFlag = false;             // 用来判断有无文件添加
     private boolean hasFetchedFlag = false;           // 用来判断有无文件取回
+    private boolean adSettingFlag = false;
+
+    private static int[] currentVideoFileClickedFlag;        // 用来判断当前播放的文件
+
+    //private File root = new File(FROMPATH);
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_ad_setting);
 
+        //Toast.makeText(AdSettingActivity.this, "请插入U盘", Toast.LENGTH_LONG);
+
         Button buttonAdReturn = (Button) findViewById(R.id.btn_ad_return);
         Button buttonAdAdd = (Button) findViewById(R.id.btn_ad_add);
         Button buttonAdDelete = (Button) findViewById(R.id.btn_ad_delete);
         Button buttonAdFetch = (Button) findViewById(R.id.btn_ad_fetch);
         Button buttonAdSetting = (Button) findViewById(R.id.btn_ad_setting);
+        Button buttonAdLoadAd = (Button) findViewById(R.id.btn_ad_load_ad);
 
         imageViewAd = (ImageView) findViewById(R.id.iv_ad);
 
@@ -61,9 +75,36 @@ public class AdSettingActivity extends Activity {
 
         listViewAdUSB = (ListView) findViewById(R.id.lv_ad_usb);
         listViewAdInternal = (ListView) findViewById(R.id.lv_ad_internal);
+        listViewAdSetting = (ListView) findViewById(R.id.lv_ad_setting);
+
+        //currentVideoFileClickedFlag = new int[2];
 
         imageViewAd.setVisibility(View.VISIBLE);
         videoViewAd.setVisibility(View.INVISIBLE);
+
+
+
+        mySettingListAdapter = new MySettingListAdapter(this, getList(TOPATH));
+        listViewAdSetting.setAdapter(mySettingListAdapter);
+        listViewAdSetting.setVisibility(View.INVISIBLE);
+
+//        if (!USBExists()) {
+//            Toast.makeText(AdSettingActivity.this, "请插入U盘", Toast.LENGTH_LONG).show();
+//        } else {
+//            myUSBListAdapter = new MyUSBListAdapter(this, getList(FROMPATH));
+//            listViewAdUSB.setAdapter(myUSBListAdapter);
+//            listViewAdUSB.setOnItemClickListener(new ListView.OnItemClickListener() {
+//                @Override
+//                public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+//                    File[] files = FileService.getFiles(FROMPATH);
+//                    playVideo(files[position].getPath());
+////                currentVideoFileClickedFlag[0] = FROMPATHFLAG;
+////                currentVideoFileClickedFlag[1] = position;
+//                    videoViewAd.setVisibility(View.VISIBLE);
+//                    imageViewAd.setVisibility(View.INVISIBLE);
+//                }
+//            });
+//        }
 
 
         myUSBListAdapter = new MyUSBListAdapter(this, getList(FROMPATH));
@@ -71,20 +112,28 @@ public class AdSettingActivity extends Activity {
         listViewAdUSB.setOnItemClickListener(new ListView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                File[] files = FileService.getFiles(FROMPATH);
-                playVideo(files[position].getPath());
-                videoViewAd.setVisibility(View.VISIBLE);
-                imageViewAd.setVisibility(View.INVISIBLE);
+                if (USBExists()) {
+                    File[] files = FileService.getFiles(FROMPATH);
+                    playVideo(files[position].getPath());
+//                currentVideoFileClickedFlag[0] = FROMPATHFLAG;
+//                currentVideoFileClickedFlag[1] = position;
+                    videoViewAd.setVisibility(View.VISIBLE);
+                    imageViewAd.setVisibility(View.INVISIBLE);
+                }
             }
         });
 
+
         myInternalListAdapter = new MyInternalListAdapter(this, getList(TOPATH));
         listViewAdInternal.setAdapter(myInternalListAdapter);
+        listViewAdInternal.setVisibility(View.VISIBLE);
         listViewAdInternal.setOnItemClickListener(new ListView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                 File[] files = FileService.getFiles(TOPATH);
                 playVideo(files[position].getPath());
+//                currentVideoFileClickedFlag[0] = TOPATHFLAG;
+//                currentVideoFileClickedFlag[1] = position;
                 videoViewAd.setVisibility(View.VISIBLE);
                 imageViewAd.setVisibility(View.INVISIBLE);
             }
@@ -94,29 +143,30 @@ public class AdSettingActivity extends Activity {
             @Override
             public void onCompletion(MediaPlayer mediaPlayer) {
                 videoViewAd.start();
+                //rePlayVideo(currentVideoFileClickedFlag[0], currentVideoFileClickedFlag[1]);
             }
         });
 
         buttonAdAdd.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-
-                File[] currentFiles = FileService.getFiles(FROMPATH);
-                for (int i = 0; i < MyUSBListAdapter.isFileAdded.length; i++) {
-                    if (MyUSBListAdapter.isFileAdded[i]) {
-                        FileService.copyFile(currentFiles[i].getPath(), TOPATH + currentFiles[i].getName());
-                        hasAddedFlag = true;
+                if (USBExists()) {
+                    File[] currentFiles = FileService.getFiles(FROMPATH);
+                    for (int i = 0; i < MyUSBListAdapter.isFileAdded.length; i++) {
+                        if (MyUSBListAdapter.isFileAdded[i]) {
+                            FileService.copyFile(currentFiles[i].getPath(), TOPATH + currentFiles[i].getName());
+                            hasAddedFlag = true;
+                        }
                     }
-                }
-
-                if (hasAddedFlag) {
-                    MyInternalListAdapter.setList(getList(TOPATH));
-                    MyInternalListAdapter.setIsSelectedAndIsFileAdded(getList(TOPATH));
-                    myInternalListAdapter.notifyDataSetChanged();
-                    //MyUSBListAdapter.setList(getList(FROMPATH));
-                    MyUSBListAdapter.setIsSelectedAndIsFileAdded(getList(FROMPATH));
-                    myUSBListAdapter.notifyDataSetChanged();
-                    hasAddedFlag = false;
+                    if (hasAddedFlag) {
+                        MyInternalListAdapter.setList(getList(TOPATH));
+                        MyInternalListAdapter.setIsSelectedAndIsFileAdded(getList(TOPATH));
+                        myInternalListAdapter.notifyDataSetChanged();
+                        //MyUSBListAdapter.setList(getList(FROMPATH));
+                        MyUSBListAdapter.setIsSelectedAndIsFileAdded(getList(FROMPATH));
+                        myUSBListAdapter.notifyDataSetChanged();
+                        hasAddedFlag = false;
+                    }
                 }
             }
         });
@@ -124,22 +174,23 @@ public class AdSettingActivity extends Activity {
         buttonAdFetch.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                File[] currentFiles = FileService.getFiles(TOPATH);
-                for (int i = 0; i < MyInternalListAdapter.isFileAdded.length; i++) {
-                    if (MyInternalListAdapter.isFileAdded[i]) {
-                        FileService.copyFile(currentFiles[i].getPath(), FROMPATH + currentFiles[i].getName());
-                        hasFetchedFlag = true;
+                if (USBExists()) {
+                    File[] currentFiles = FileService.getFiles(TOPATH);
+                    for (int i = 0; i < MyInternalListAdapter.isFileAdded.length; i++) {
+                        if (MyInternalListAdapter.isFileAdded[i]) {
+                            FileService.copyFile(currentFiles[i].getPath(), FROMPATH + currentFiles[i].getName());
+                            hasFetchedFlag = true;
+                        }
                     }
-                }
-
-                if (hasFetchedFlag) {
-                    MyUSBListAdapter.setList(getList(FROMPATH));
-                    MyUSBListAdapter.setIsSelectedAndIsFileAdded(getList(FROMPATH));
-                    myUSBListAdapter.notifyDataSetChanged();
-                    //MyInternalListAdapter.setList(getList(FROMPATH));
-                    MyInternalListAdapter.setIsSelectedAndIsFileAdded(getList(TOPATH));
-                    myInternalListAdapter.notifyDataSetChanged();
-                    hasFetchedFlag = false;
+                    if (hasFetchedFlag) {
+                        MyUSBListAdapter.setList(getList(FROMPATH));
+                        MyUSBListAdapter.setIsSelectedAndIsFileAdded(getList(FROMPATH));
+                        myUSBListAdapter.notifyDataSetChanged();
+                        //MyInternalListAdapter.setList(getList(FROMPATH));
+                        MyInternalListAdapter.setIsSelectedAndIsFileAdded(getList(TOPATH));
+                        myInternalListAdapter.notifyDataSetChanged();
+                        hasFetchedFlag = false;
+                    }
                 }
             }
         });
@@ -147,33 +198,88 @@ public class AdSettingActivity extends Activity {
         buttonAdDelete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                if (FileService.isFileAdded(FROMPATH, MyUSBListAdapter.isFileAdded)
-                        || FileService.isFileAdded(TOPATH, MyInternalListAdapter.isFileAdded)) {
-                    AlertDialog.Builder builder = new AlertDialog.Builder(AdSettingActivity.this);
-                    builder.setMessage("确认删除吗？");
-                    builder.setTitle("提示");
-                    builder.setPositiveButton("确认", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            dialog.dismiss();
-                            FileService.deleteFile(FROMPATH, MyUSBListAdapter.isFileAdded);
-                            FileService.deleteFile(TOPATH, MyInternalListAdapter.isFileAdded);
+                if (USBExists()) {
+                    if (FileService.isFileAdded(FROMPATH, MyUSBListAdapter.isFileAdded)
+                            || FileService.isFileAdded(TOPATH, MyInternalListAdapter.isFileAdded)) {
+                        AlertDialog.Builder builder = new AlertDialog.Builder(AdSettingActivity.this);
+                        builder.setMessage("确认删除吗？");
+                        builder.setTitle("提示");
+                        builder.setPositiveButton("确认", new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface dialog, int which) {
+                                dialog.dismiss();
+                                FileService.deleteFile(FROMPATH, MyUSBListAdapter.isFileAdded);
+                                FileService.deleteFile(TOPATH, MyInternalListAdapter.isFileAdded);
 
-                            MyInternalListAdapter.setList(getList(TOPATH));
-                            MyInternalListAdapter.setIsSelectedAndIsFileAdded(getList(TOPATH));
-                            myInternalListAdapter.notifyDataSetChanged();
-                            MyUSBListAdapter.setList(getList(FROMPATH));
-                            MyUSBListAdapter.setIsSelectedAndIsFileAdded(getList(FROMPATH));
-                            myUSBListAdapter.notifyDataSetChanged();
-                        }
-                    });
-                    builder.setNegativeButton("取消", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            dialog.dismiss();
-                        }
-                    });
-                    builder.create().show();
+                                MyInternalListAdapter.setList(getList(TOPATH));
+                                MyInternalListAdapter.setIsSelectedAndIsFileAdded(getList(TOPATH));
+                                myInternalListAdapter.notifyDataSetChanged();
+                                MyUSBListAdapter.setList(getList(FROMPATH));
+                                MyUSBListAdapter.setIsSelectedAndIsFileAdded(getList(FROMPATH));
+                                myUSBListAdapter.notifyDataSetChanged();
+                            }
+                        });
+                        builder.setNegativeButton("取消", new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface dialog, int which) {
+                                dialog.dismiss();
+                            }
+                        });
+                        builder.setCancelable(false);
+                        builder.create().show();
+                    }
+                } else {
+                    if (FileService.isFileAdded(TOPATH, MyInternalListAdapter.isFileAdded)) {
+                        AlertDialog.Builder builder = new AlertDialog.Builder(AdSettingActivity.this);
+                        builder.setMessage("确认删除吗？");
+                        builder.setTitle("提示");
+                        builder.setPositiveButton("确认", new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface dialog, int which) {
+                                dialog.dismiss();
+                                FileService.deleteFile(TOPATH, MyInternalListAdapter.isFileAdded);
+                                MyInternalListAdapter.setList(getList(TOPATH));
+                                MyInternalListAdapter.setIsSelectedAndIsFileAdded(getList(TOPATH));
+                                myInternalListAdapter.notifyDataSetChanged();
+                            }
+                        });
+                        builder.setNegativeButton("取消", new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface dialog, int which) {
+                                dialog.dismiss();
+                            }
+                        });
+                        builder.setCancelable(false);
+                        builder.create().show();
+                    }
+                }
+            }
+        });
+
+        buttonAdSetting.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                if (adSettingFlag) {
+                    listViewAdSetting.setVisibility(View.INVISIBLE);
+                    listViewAdInternal.setVisibility(View.VISIBLE);
+                    adSettingFlag = false;
+                } else {
+                    listViewAdInternal.setVisibility(View.INVISIBLE);
+                    listViewAdSetting.setVisibility(View.VISIBLE);
+                    adSettingFlag = true;
+                }
+            }
+        });
+
+        buttonAdLoadAd.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                if (USBExists()) {
+                    MyUSBListAdapter.setList(getList(FROMPATH));
+                    MyUSBListAdapter.setIsSelectedAndIsFileAdded(getList(FROMPATH));
+                    myUSBListAdapter.notifyDataSetChanged();
+                } else {
+                    Toast.makeText(AdSettingActivity.this, "请插入U盘", Toast.LENGTH_LONG).show();
                 }
             }
         });
@@ -188,22 +294,59 @@ public class AdSettingActivity extends Activity {
             imageViewAd.setVisibility(View.VISIBLE);
             videoViewAd.setVisibility(View.INVISIBLE);
         }
-
     }
+
+//    public void rePlayVideo(int filePath, int position) {
+//        switch (filePath) {
+//            case FROMPATHFLAG:
+//                File[] fromFiles = FileService.getFiles(FROMPATH);
+//                playVideo(fromFiles[position].getPath());
+//                break;
+//            case TOPATHFLAG:
+//                File[] toFiles = FileService.getFiles(TOPATH);
+//                playVideo(toFiles[position].getPath());
+//                break;
+//        }
+//    }
 
     private List<Map<String, Object>> getList(String filePath)
     {
-        File[] files = FileService.getFiles(filePath);
-        List<Map<String, Object>> list = new ArrayList<Map<String, Object>>();
-        Map<String, Object> map;
-        for(int i = 0; i < files.length; i++)
-        {
-            map = new HashMap<String, Object>();
-            map.put("img", R.mipmap.ic_launcher);
-            map.put("title", files[i].getName());
-            list.add(map);
+        File file = new File(filePath);
+
+        if (!file.exists()) {
+            //File[] files = FileService.getFiles(filePath);
+            List<Map<String, Object>> list = new ArrayList<Map<String, Object>>();
+            Map<String, Object> map;
+            for(int i = 0; i < 10; i++)
+            {
+                map = new HashMap<String, Object>();
+                map.put("img", R.mipmap.ic_launcher);
+                map.put("title", "请插入U盘");
+                list.add(map);
+            }
+            return list;
+        } else {
+            File[] files = FileService.getFiles(filePath);
+            List<Map<String, Object>> list = new ArrayList<Map<String, Object>>();
+            Map<String, Object> map;
+            for(int i = 0; i < files.length; i++)
+            {
+                map = new HashMap<String, Object>();
+                map.put("img", R.mipmap.ic_launcher);
+                map.put("title", files[i].getName());
+                list.add(map);
+            }
+            return list;
         }
-        return list;
+    }
+
+    private boolean USBExists() {
+        File usbFile = new File(FROMPATH);
+        if (usbFile.exists()) {
+            return true;
+        } else {
+            return false;
+        }
     }
